angle参数类型更新 float-->int

+++ stupidStupid/stupidStupid.h
@@ -7,7 +7,7 @@ typedef struct{
   
   int speedMax = 0;
   float speedRate = 0.0;//速度比例
-  float angleValue = 0.0;//角度值
+  int angleValue = 0;//角度值
 
   float spinSpeedRate = 0.0;//旋转比例
 }sMatDat_t;
